requirement.txt
numpy==2.2.4
packaging==24.2
pandas==2.2.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.56.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
nltk==3.9.1
pillow==11.1.0
pyaml==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scikit-optimize==0.10.2
scipy==1.15.2
six==1.17.0
sklearn-compat==0.1.3
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
zxcvbn==4.5.0
lightgbm==4.6.0
